+++ packages/cli/tests/__data__/cli-graphql-doc-generator-groups.config.js
@@ -31,7 +31,6 @@ module.exports = {
         docOptions: {
           index: true,
           categorySort: "natural",
-          categorySortPrefix: true,
         },
         customDirective: {
           beta: {