+++ packages/docusaurus/tests/__data__/docusaurus2-graphql-doc-build.js
@@ -49,7 +49,6 @@ module.exports = {
         runOnBuild: true,
         docOptions: {
           categorySort: "natural",
-          categorySortPrefix: true,
         },
         printTypeOptions: {
           hierarchy: "api",