+++ packages/docusaurus/tests/__data__/docusaurus2-graphql-doc-generator-groups.config.js
@@ -28,7 +28,6 @@ module.exports = {
   docOptions: {
     index: true,
     categorySort: "natural",
-    categorySortPrefix: true,
   },
   customDirective: {
     beta: {