+++ packages/docusaurus/tests/__data__/docusaurus2-graphql-doc-nobuild.js
@@ -49,7 +49,6 @@ module.exports = {
         runOnBuild: false,
         docOptions: {
           categorySort: "natural",
-          categorySortPrefix: true,
         },
       },
     ],